Pin the FSFS f7 item type to a fixed-length data type. Make sure we use that type consistently.
* subversion/libsvn_fs_fs/index.h
  (svn_fs_fs__p2l_entry_t): Fix the item type to 32 bits which is way
                            more than we'll ever need and which aligns
                            nicely with the rest of the struct.

* subversion/libsvn_fs_fs/index.c
  (read_entry): Remove unnecessary type conversion.

* subversion/libsvn_fs_fs/cached_data.c
  (dbg_log_access): Always use uint32 to hold item types.

* subversion/libsvn_fs_fs/transaction.c
  (write_container_rep,
   write_container_delta_rep,
   write_final_rev): Same.

git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1627966 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/libsvn_fs_fs/index.c
@@ -2068,7 +2068,7 @@ read_entry(svn_fs_fs__packed_number_stream_t *stream,
   SVN_ERR(packed_stream_get(&value, stream));
   *last_compound += decode_int(value);
 
-  entry.type = (int)(*last_compound & 7);
+  entry.type = *last_compound & 7;
   entry.item.number = *last_compound / 8;
 
   /* Verify item type. */
